refactor: Wave 1 - relocate loose root files to Core/
- SFRCO.cpp/.h -> Core/Net/ (the shared networking seam; sets up Wave 2's Net/ extraction)
- SF_ATAnchor.h -> Core/
- rename SFHologramHelperService_Part2.cpp -> _Children.cpp
  (concern: child-hologram spawn / reposition / lock / destroy)
- module identity + logging core (SmartFoundations.h/.cpp, SFLogMacros.h) stay at the module
  root by UE convention (~87 includers); recorded as a policy exception
- all includes rewritten (one used the non-standard "Public/SFRCO.h" form - caught by the
  compile gate); Shipping module links clean

Per docs/Development/CodeOrganization.md.
