+++ Source/SmartFoundations/Private/Features/Extend/SFExtendService.cpp
@@ -23,7 +23,7 @@
 #include "Subsystem/SFSubsystem.h"
 #include "Subsystem/SFHologramDataService.h"
 #include "Core/Helpers/SFNetworkHelper.h"   // [EXTEND-MP] IsClient (server-walk topology branch)
-#include "SFRCO.h"                          // [EXTEND-MP] Server_RequestExtendTopology
+#include "Core/Net/SFRCO.h"                          // [EXTEND-MP] Server_RequestExtendTopology
 #include "FGPlayerController.h"
 // NOTE: SFRecipeCostInjector.h removed - child holograms automatically aggregate costs via GetCost()
 #include "Services/RadarPulse/SFRadarPulseService.h"
