+++ Source/SmartFoundations/Private/SmartFoundations.cpp
@@ -4,7 +4,7 @@
 #include "SFLogMacros.h"
 #include "Input/SFInputRegistry.h"
 #include "Logging/SFLogRegistry.h"
-#include "SFRCO.h"
+#include "Core/Net/SFRCO.h"
 #include "Engine/Engine.h"
 #include "Registry/RemoteCallObjectRegistry.h"
